tests/wpt/web-platform-tests/css/css-fonts/parsing/font-size-adjust-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 3: parsing font-size-adjust with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-size-adjust-prop">
<meta name="assert" content="font-size-adjust supports only the grammar 'none | <number>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value('font-size-adjust', 'auto');
test_invalid_value('font-size-adjust', '-10');
</script>
</body>
</html>
